+++ frb_codegen/src/library/codegen/generator/api_dart/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod internal_config;
+pub(crate) mod misc;
 pub(crate) mod spec_generator;
-pub(crate) mod text_generator;
+mod text_generator;
 
 use crate::codegen::dumper::internal_config::ConfigDumpContent;
 use crate::codegen::dumper::Dumper;
